requirements.txt
streamlit
pandas
supabase
python-dotenv
streamlit-option-menu
streamlit-ace
google-genai